refactor: extract valida/createDataLoaders functions out of `index.ts`
